# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: WASAPhoto API
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to 
    WASAPhoto! You can upload your photos directly from your PC, and they will be 
    to everyone following you.
  version: 1.0.0
...
components:
  securitySchemes:
    # Bearer authentication using the user identifier instead of the token
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token containing the userID.
  schemas:
    UserID:
      description: user ID
      type: integer
      example: 1234
    Username:
      description: username
      type: string
      example: Maria
      pattern: '^[a-zA-Z0-9]+$'
      minLength: 3
      maxLength: 16
    PhotoID:
      type: integer
      description: ID of the photo.
      example: 1234
...
    CommentID:
      description: comment ID
      type: integer
      example: 1234
    ImageData:
      type: string
      description: Binary data representing the image.
      format: binary
      minLength: 1
      maxLength: 5000
    Photo:
      description: |- 
        Photo object that includes the author's name, the image URL, the number of
        likes and comments, and details about users who liked or commented, including
        the comments themselves.
      type: object
      properties:
      
        author:
          $ref: '#/components/schemas/Username'
        
        photoID:
          $ref: '#/components/schemas/PhotoID'
          
        imageData:
          $ref: '#/components/schemas/ImageData'
      
        uploadDate:
          type: string
          description: The date and time when the upload occurred.
          format: date-time
          example: 2023-11-09T15:30:00Z
          minLength: 1
          maxLength: 35
          
        likesCount:
          type: integer
          description: number of likes
          example: 12
          
        whoLiked:
          type: array
          minItems: 0
          maxItems: 5000
          description: Contains the user IDs of all users who have liked.
          items:
            $ref: '#/components/schemas/UserID'
            
        commentsCount:
          type: integer
          description: number of comments
          example: 3
          
        comments:
          type: array
          description: |-
            Contains the user IDs of all users who have commented along with their
            respective comments.
          minItems: 0
          maxItems: 5000
          items:
            type: object
            properties:
              userID:
                $ref: '#/components/schemas/UserID'
              commentID:
                $ref: '#/components/schemas/CommentID'
              comment:
                type: string
                description: Comment on a photo
                pattern: '^.*$'
                minLength: 1
                maxLength: 5000
                example: Beautiful photo!
  responses:
    UnauthorizedError:
      description: user ID is missing or invalid
      content:
        application/json:
          schema:
            description: Contains an error message
            type: object
            properties:
              message:
                description: error message
                type: string
                example: Invalid user ID supplied
    InvalidSyntax:
      description: Invalid syntax
      content:
        application/json:
          schema:
            description: Contains an error message
            type: object
            properties:
              message:
                description: error message
                type: string
                example: Invalid username format. Please follow the specified requirements.
    InvalidRequest: 
      description: Invalid request
      content:
        application/json:
          schema:
            description: Contains an error message
            type: object
            properties:
              message:
                description: error message
                type: string
                example: Invalid request
    NotFoundError:
      description: Element not found
      content:
        application/json:
          schema: 
            description: Contains an error message
            type: object
            properties:
              message:
                description: error message
                type: string
                example: Element not found
security:
  - bearerAuth: []
paths:
...
  /users/{userID}:
    parameters:
      - name: userID
        in: path
        required: true
        description: ID of the user
        schema:
          $ref: '#/components/schemas/UserID'
          
    put:
      tags: ["User"]
      summary: Sets the username
      description: The user can change their username.
      operationId: setMyUserName
      requestBody:
        description: User's username
        required: true
        content:
          application/json:
            schema:
              description: Contains the username
              type: object
              properties:
                name:
                  $ref: '#/components/schemas/Username'
                  
      responses:
        '200':
          description: Username changed successfully
          content: 
            application/json:
              schema:
                description: Contains a success message
                type: object
                properties:
                  message:
                    description: OK message
                    type: string
                    example: OK
        
        '400':
          $ref: '#/components/responses/InvalidSyntax'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'

    get:
      tags: ["User"]
      summary: View the profile of an user
      description: |-
        The user's personal profile page displays their photos and the number of
        followers, following and uploaded photos. A user can search other user
        profiles via username.
      operationId: getUserProfile
      
      responses:
        '200':
          description: OK
          content:
            multipart/form-data:
              schema:
                description: returns the array containing all the photos and details
                type: object
                properties:
                  photos:
                    type: array
                    description: Contains all the photos and details
                    minItems: 0
                    maxItems: 5000
                    items:
                      $ref: '#/components/schemas/Photo'
                      
                  
            application/json:
              schema:
                description: |-
                  Contains the number of photos posted, followers and
                  following
                type: object
                properties:
                  photosCount:
                    type: integer
                    description: Number of photos posted by the user
                  followersCount:
                    type: integer
                    description: Number of followers
                  followingCount:
                    type: integer
                    description: Number of users being followed
          
        '400':
          $ref: '#/components/responses/InvalidRequest'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
          
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/{userID}/photos:
    post:
      tags: ["User","Photos"]
      summary: Uploads a photo
      description: |- 
        The user can post a new photo on their profile. Photos will be presented in
        reverse chronological order.
      operationId: uploadPhoto
      requestBody:
        description: URL of the photo that the user wants to upload.
        required: true
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/ImageData'
      parameters:
        - name: userID
          in: path
          required: true
          description: ID of the user.
          schema:
            $ref: '#/components/schemas/UserID'
      
      responses:
        '201':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
  /users/{userID}/following:
    put:
      tags: ["User"]
      summary: Adds a user to the following list of the specified user
      description: Users can follow others to view and interact with their photos.
      operationId: followUser
      parameters:
        - name: userID
          in: path
          required: true
          description: ID of the user.
          schema:
            $ref: '#/components/schemas/UserID'
      requestBody:
        description: ID of the user you want to follow.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserID'
            
                    
      responses:
        '201':
          description: User followed successfully
          content:
            application/json:
              schema:
                description: Contains a success message
                properties:
                  message:
                    description: OK message
                    type: string
                    example: OK
                  
          
        '400':
          $ref: '#/components/responses/InvalidRequest'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
          
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/{userID}/following/{followingID}:
    delete:
      tags: ["User"]
      summary: Removes a user from the following list of the specified user
      description: |-
        Users can unfollow someone they were previously following to stop seeing
        their photos.
      operationId: unfollowUser
      parameters:
        - name: userID
          in: path
          required: true
          description: ID of the user.
          schema:
            $ref: '#/components/schemas/UserID'
        - name: followingID
          in: path
          required: true
          description: ID of the user you want to unfollow.
          schema:
            $ref: '#/components/schemas/UserID'
        
        
                    
      responses:
        '201':
          description: User unfollowed successfully
          content: 
            application/json:
              schema:
                description: Contains a success message
                type: object
                properties:
                  message:
                    description: OK message
                    type: string
                    example: OK
          
        '400':
          $ref: '#/components/responses/InvalidRequest'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
          
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/{userID}/banned_users:
    put:
      tags: ["User"]
      summary: Adds a user to the list of users banned by the specified user
      description: |-
        Users can block other one to prevent them from seeing any information
        about them.
      operationId: banUser
      parameters:
        - name: userID
          in: path
          required: true
          description: ID of the user.
          schema:
            $ref: '#/components/schemas/UserID'
      requestBody:
        description: ID of the user you want to ban.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserID'
                    
      responses:
        '201':
          description: User banned successfully
          content: 
            application/json:
              schema:
                description: Contains a success message
                type: object
                properties:
                  message:
                    description: OK message
                    type: string
                    example: OK
          
        '400':
          $ref: '#/components/responses/InvalidRequest'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
          
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/{userID}/banned_users/{bannedUserID}:
    delete:
      tags: ["User"]
      summary: Removes a user from the list of users banned by the specified user
      description: Unblocking a user allows them to view your profile information.
      operationId: unbanUser
      parameters:
        - name: userID
          in: path
          required: true
          description: ID of the user.
          schema:
            $ref: '#/components/schemas/UserID'
        - name: bannedUserID
          in: path
          required: true
          description: ID of the user you want to unban.
          schema:
            $ref: '#/components/schemas/UserID'
                    
      responses:
        '201':
          description: User unbanned successfully
          content: 
            application/json:
              schema:
                description: Contains a success message
                type: object
                properties:
                  message:
                    description: OK message
                    type: string
                    example: OK
          
        '400':
          $ref: '#/components/responses/InvalidRequest'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
          
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/{userID}/stream:
    get:
      tags: ["User"]
      summary: View the user's stream
      description: |-
        The stream is composed by photos from “following” in reverse chronological
        order.
      operationId: getMyStream
      parameters:
        - name: userID
          in: path
          required: true
          description: ID of the user.
          schema:
            $ref: '#/components/schemas/UserID'
        
      responses:
        '200':
          description: OK
          content:
            multipart/form-data:
              schema:
                description: returns the array containing all the photos and details
                type: object
                properties:
                  photos:
                    type: array
                    description: Contains all the photos and details
                    minItems: 0
                    maxItems: 5000
                    items:
                      $ref: '#/components/schemas/Photo'
          
        '400':
          $ref: '#/components/responses/InvalidRequest'
          
        '401': 
          $ref: '#/components/responses/UnauthorizedError'
...